Extract view-in-browser logic to a controller
[MAILPOET-2430]
